fix update item

+++ lib/core/widgets/forms/RoundedNumberField.dart
@@ -9,7 +9,7 @@ class RoundedNumberField extends StatelessWidget {
   final FormFieldValidator<String>  validator;
   final TextEditingController controller;
   final ValueChanged<String> onChanged;
-  final int initialValue;
+  final double initialValue;
 
   const RoundedNumberField({Key key, this.hintText, this.validator, this.controller, this.onChanged, this.initialValue}) : super(key: key);
